[Sockets] Adds sale-wide event handling for current lot changes.
