+++ Artsy/View_Controllers/Live_Auctions/Views/LiveAuctionCurrentLotView.swift
@@ -1,0 +1,66 @@
+import UIKit
+import Interstellar
+
+
+class LiveAuctionCurrentLotView: UIButton {
+
+    let viewModel = Observable<LiveAuctionLotViewModelType>()
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+
+        backgroundColor = .artsyPurpleRegular()
+
+        let liveLotLabel = ARSansSerifLabel()
+        liveLotLabel.font = .sansSerifFontWithSize(12)
+        liveLotLabel.text = "Live Lot"
+
+        let artistNameLabel = UILabel()
+        artistNameLabel.font = .serifSemiBoldFontWithSize(16)
+
+        let biddingPriceLabel = ARSansSerifLabel()
+        biddingPriceLabel.font = .sansSerifFontWithSize(16)
+
+        let hammerView = UIImageView(image: UIImage(asset: .Lot_bidder_hammer_white))
+        let thumbnailView = UIImageView(frame: CGRect(x: 0, y: 0, width: 40, height: 40))
+
+        [liveLotLabel, artistNameLabel, biddingPriceLabel, thumbnailView, hammerView].forEach { addSubview($0) }
+        [liveLotLabel, artistNameLabel, biddingPriceLabel].forEach {
+            $0.backgroundColor = backgroundColor
+            $0.textColor = .whiteColor()
+        }
+
+        constrainHeight("54")
+
+        // Left Side
+
+        thumbnailView.alignLeadingEdgeWithView(self, predicate: "10")
+        thumbnailView.constrainWidth("38", height: "38")
+        thumbnailView.alignCenterYWithView(self, predicate: "0")
+
+        liveLotLabel.constrainLeadingSpaceToView(thumbnailView, predicate: "10")
+        liveLotLabel.alignTopEdgeWithView(self, predicate: "10")
+
+        artistNameLabel.constrainLeadingSpaceToView(thumbnailView, predicate: "10")
+        artistNameLabel.alignBottomEdgeWithView(self, predicate: "-10")
+
+        // Right side
+
+        hammerView.alignTrailingEdgeWithView(self, predicate: "-10")
+        hammerView.constrainWidth("32", height: "32")
+        hammerView.alignCenterYWithView(self, predicate: "0")
+
+        biddingPriceLabel.alignAttribute(.Trailing, toAttribute: .Leading, ofView: hammerView, predicate: "-12")
+        biddingPriceLabel.alignCenterYWithView(self, predicate: "0")
+
+        viewModel.subscribe { vm in
+            artistNameLabel.text = vm.lotArtist
+            biddingPriceLabel.text = vm.currentLotValueString
+            thumbnailView.ar_setImageWithURL(vm.urlForThumbnail)
+        }
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
