+++ CHANGELOG.yml
@@ -31,6 +31,7 @@ upcoming:
     - Adds implementation to handle some socket events. - ash
     - Adds more socket event-handling. - ash
     - Staging causality socket points to staging now instead of localhost. - ash
+    - Adds sale-wide event handling (changes to current lot). - ash
 
   notes:
     - Fixes long auction names falling off right edge. - ash
